rename: 데이터 초기 세팅하는 파일들 config패키지에서 init패키지로 이동
