+++ src/main/java/com/soda/global/config/AWSConfig.java
@@ -1,0 +1,4 @@
+package com.soda.global.config;
+
+public class AWSConfig {
+}
