+++ src/main/java/com/soda/global/init/InitialDataLoader.java
@@ -1,6 +1,5 @@
-package com.soda.global.config;
+package com.soda.global.init;
 
-import com.soda.member.entity.Company;
 import com.soda.member.entity.Member;
 import com.soda.member.enums.MemberRole;
 import com.soda.member.repository.CompanyRepository;
